Preserve report when test fails
- Report was being lost because of how the cleanup
  logic works.  It tries to check the call stack to
  see if an exception is present (e.g. for a failed
  test).  Since all tests are run as a module, if
  one test fails, but a subsequent test succeeds
  the logic to cleanup files is activated because
  the original exception from the failed test is
  lost.
- Moved long test decorator up to class since the
  run is performed in class setUp.

+++ test_integration/test_ee_stream_dgemm_mix.py
@@ -50,6 +50,7 @@ from test_integration import geopm_test_launcher
 from test_integration import util
 
 _g_skip_launch = False
+_g_keep_reports = False
 
 class AppConf(object):
     """Class that is used by the test launcher as a geopmpy.io.BenchConf
@@ -75,6 +76,7 @@ class AppConf(object):
 
 @util.skip_unless_cpufreq()
 @util.skip_unless_optimized()
+@util.skip_unless_run_long_tests()
 class TestIntegrationEEStreamDGEMMMix(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
@@ -114,13 +116,19 @@ class TestIntegrationEEStreamDGEMMMix(unittest.TestCase):
 
         """
         if (sys.exc_info() == (None, None, None) and not
-            cls._keep_files and not cls._skip_launch):
+            cls._keep_files and not cls._skip_launch and not
+            _g_keep_reports):
             os.unlink(cls._agent_conf_path)
             os.unlink(cls._report_path)
             for ff in glob.glob(cls._trace_path + '*'):
                 os.unlink(ff)
 
-    @util.skip_unless_run_long_tests()
+    def tearDown(self):
+        """Set keep_reports global for any failed test."""
+        if sys.exc_info() != (None, None, None):
+            global _g_keep_reports
+            _g_keep_reports = True
+
     def test_monotone_frequency(self):
         """Test that agent selects lower frequencies for regions with more
         stream.
